Network is on server side, training only once, when starting the server
