+++ public/index.html
@@ -13,8 +13,6 @@
     	<script src="https://cdn.socket.io/socket.io-1.4.5.js"></script>
 		
 		<script src="sketch.js"></script>
-		<script src="NeuralNetwork.js"></script>
-		<script src="MatrixLib.js"></script>
 	</head>
 	<body style="background-color: black;">
 	</body>
